chore: reorganise server tests by test type
