+++ clinic-system/server/tests/unit/validation/patientValidation.test.js
@@ -2,7 +2,7 @@ const {
   isValidUuid,
   isValidEmail,
   validatePatientInput,
-} = require('..//src/patientValidation')
+} = require('../../../src/patientValidation')
 
 describe('patientValidation', () => {
   const validUuid = '123e4567-e89b-12d3-a456-426614174000'
